Remove some logic from template into grunt-browser-db-task
Adding legacy flag and human readable version.
