+++ tasks/db/grunt-browserhacks-db-task.js
@@ -1,5 +1,7 @@
-var addId   = require('./addid'),
-    cssLint = require('./csslint');
+var addId      = require('./addid'),
+    cssLint    = require('./csslint'),
+    addVersion = require('./formatVersion.js'),
+    addLegacy   = require('./isLegacy');
 
 module.exports = function(grunt) {
   grunt.registerTask('updateDatabase', 'Update the database used by Browserhacks.', function(){
@@ -14,11 +16,16 @@ module.exports = function(grunt) {
     }
 
     var hacks       = grunt.file.readJSON(src),
+        browsers    = grunt.file.readJSON('./src/db/browsers.json'), // @TODO Make this customizable in Gruntfile.js
         typeArrays  = {};
-
 
     hacks.forEach(function(hack){
       addId(hack);
+
+      addVersion(hack, browsers);
+
+
+      addLegacy(hack, browsers);
       cssLint(hack);
       //@TODO isLegacy(hack);
